% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAllCorHats.R
\name{computeAllCorHats}
\alias{computeAllCorHats}
\alias{computeAllRhoHats}
\title{computeAllCorHats}
\usage{
computeAllCorHats(tParArray, lagCount, sample)
}
\arguments{
\item{tParArray}{a T parameter array, a.k.a. a sleeper parameter, a sequence of numbers {0, 1}.}

\item{lagCount}{lag.}

\item{sample}{sample.}

\item{kernel}{kernel.}

\item{bandwidth}{bandwidth.}
}
\value{

}
\description{
Computing correlations hat, that is, all \eqn{\rho} hats.
}
\examples{
tParCount = 10
tParArray <- createTParArray(tParCount = tParCount)
tvMA1CoefArray <- createTVMA1CoefArray(coefFunction = sin, tParArray = tParArray)
noise <- createNoise(sampleSize = tParCount, mean = 0, sd = 1)
mockTVMA1Array <- createTVMA1CoefArray(coefFunction = sin, tParArray = tParArray)
sample <- createSample(model = createTVMA1, tvMA1CoefArray = tvMA1Array, noise = noise)
lagCount = tParCount - 1
kernel = normalDifferenceKernel
allCorHats = computeAllCorHats(tParArray = tParArray, lagCount = lagCount, sample = sample, kernel = kernel,  bandwidth = 1)
}
