% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBetaLRVHat.R
\name{computeBetaLRVHat}
\alias{computeBetaLRVHat}
\alias{computeGammaHat}
\title{computeBetaLRVHat}
\usage{
computeBetaLRVHat(tParArray, lag, sample, allCorHats)
}
\arguments{
\item{tParArray}{a T parameter array, a.k.a. a sleeper parameter, a sequence of numbers {0, 1}.}

\item{lag}{a lag value.}

\item{sample}{a sample.}

\item{bandwidth}{by default, bandwidth = 1.}
}
\value{
A scalar value of a \code{Beta LRV Hat} given certain \code{lag}.
}
\description{
Computing a \code{betaLRVHat} given certain \code{lag}.
}
\examples{
tParCount = 10
tParArray <- createTParArray(tParCount = tParCount)
noise <- createNoise(sampleSize = tParCount, mean = 0, sd = 1)
tvMA1CoefArray <- createTVMA1CoefArray(coefFunction = sin, tParArray = tParArray)
sample <- createSample(model = createMa1, tvMA1CoefArray = tvMA1CoefArray, noise = noise)
bandwith = 1
lag = 2
allCorHats = computeAllCorHats(tParArray = tParArray, lagCount = tParCount - 1, sample = sample, bandwidth = bandwidth)
betaLRVHat = computeBetaLRVHat(tParArray = tParArray, lag = lag, sample = sample, bandwidth = bandwidth, allCorHats = allCorHats)

}
