% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCor.R
\name{computeCor}
\alias{computeCor}
\alias{computeRho}
\title{computeCor}
\usage{
computeCor(lag, tParArray)
}
\arguments{
\item{lag}{the lag}

\item{tvMa1CoefArray}{the tvMa1CoefArray}
}
\value{
Array \code{cor_lag(i)}, counter \code{i} corresponds to \code{tvMa1CoefArray_i = 0, 1, ..., length(tvMa1CoefArray)}. length(tvMa1CoefArray) = sampleSize
}
\description{
Computing correlation: \code{cor} (\eqn{\rho}) given \code{tvMa1CoefArray}.
}
\examples{
tParCount <- 10
lag <- 0
tParArray <- createTParArray(tParCount = tParCount)
tvMA1CoefArray <- createTVMA1CoefArray(coefFunction = sin, tParArray = tParArray)

}
