% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCorHat.R
\name{computeCorHat}
\alias{computeCorHat}
\alias{computeRhoHat}
\title{computeCorHat}
\usage{
computeCorHat(tParArray, lag, sample)
}
\arguments{
\item{tParArray}{a T parameter, a.k.a. a sleeper parameter, a sequence of numbers {0, 1}.}

\item{lag}{a lag value.}

\item{sample}{a sample.}

\item{bandwidth}{by default, bandwidth = 1.}
}
\value{
A scalar value of a ratio of \code{correlation hat} given certain \code{lag} over \code{correlation hat} of \code{lag = 0}
}
\description{
Computing a ratio of \code{correlation hat} given certain \code{lag} over \code{correlation hat} of \code{lag = 0}.
}
\examples{
tParCount = 10
tParArray <- createTParArray(tParCount = tParCount)
noise <- createNoise(sampleSize = tParCount, mean = 0, sd = 1)
tvMA1CoefArray <- createTVMA1CoefArray(coefFunction = sin, tParArray = tParArray)
sample <- createSample(model = createMa1, tvMA1CoefArray = tvMA1CoefArray, noise = noise)
lag = 2
corHat = computeCorHat(tParArray = tParArray, lag = lag, sample = sample, bandwidth = bandwith)

}
