% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCovHat.R
\name{computeCovHat}
\alias{computeCovHat}
\alias{}
\title{computeCovHat}
\usage{
computeCovHat(tParArray, lag, sample)
}
\arguments{
\item{tParArray}{a T parameter, a.k.a. a sleeper parameter, a sequence of numbers {0, 1}.}

\item{lag}{a lag value.}

\item{sample}{a sample.}

\item{kernel}{a kernel function.}

\item{bandwidth}{by default, bandwidth = 1.}
}
\value{

}
\description{

}
\examples{
tParArray <- createTParArray(tParCount = tParCount)
sample <- createSample(model = createMa1, tvMA1CoefArray = tvMA1CoefArray, noise = noise)
kernel  <- normalDiferencekernel
bandwith <- 1
lag <- 2

}
