% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIsCoveredArray.R
\name{computeIsCoveredArray}
\alias{computeIsCoveredArray}
\alias{}
\title{computeIsCoveredArray}
\usage{
computeIsCoveredArray(bandsBrick, corArray, tParCount, sampleSize,
  replicationCount, lag, superReplicationCount, nonCoverageProbability,
  fileName = "")
}
\arguments{
\item{bandsBrick}{bandsBrick}

\item{corArray}{corArray}
}
\value{
Brick of vectors with values {0, 1}.
}
\description{
in a cycle for each band in a brick call computeIsCovered
}
\examples{
mySampleSize=5
myTParCount = 5
mockTParArray <- createTParArray(tParCount = myTParCount)
mockTVMA1Array <- createTVMA1CoefArray(coefFunction = sin,sampleSize = mySampleSize)
myReplicationCount=2
myLag = 1
myLagCount = 3
myKernel = normalDifferenceKernel
myBandwidth = 0.5
myNonCoverageProbability=0.05
fileName <- paste("ss", mySampleSize, "l", myLag, "bandW", myBandwidth, "alpha"
               , myNonCoverageProbability, sep = "_")
bandsBrick=createBandsBrick(tParArray=mockTParArray,
                           lag=myLag,
                          lagCount=myLagCount,
                           bandwidth= myBandwidth,
                           kernel = normalDifferenceKernel,
                           sampleSize=mySampleSize,
                           nonCoverageProbability = myNonCoverageProbability,
                           replicationCount=myReplicationCount)
mockCorArray <- computeCor(lag = myLag,coefFunction = sin,tParArray = mockTParArray)
isCoveredArray <-
computeIsCoveredArray(bandsBrick = bandsBrick,
                       corArray = mockCorArray,
                      fileName = fileName)

}
