% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMEbyCovHat.R
\name{computeMEbyCovHat}
\alias{computeMEbyCovHat}
\alias{}
\title{computeMEbyCovHat}
\usage{
computeMEbyCovHat(tParArray, lag, lagCount, sample, bandwidth,
  nonCoverageProbability, allCorHats, C_K = -1.978325,
  PHI_K_NORMAL_DIFF = 0.4065)
}
\arguments{
\item{tParArray, }{lag,
lagCount,
sample,
bandwidth,
nonCoverageProbability,
allCorHats,
C_K = -1.978325,
PHI_K_NORMAL_DIFF = 0.4065}
}
\value{
Margin of error
}
\examples{
tParCount = 10
tParArray <- createTParArray(tParCount = tParCount)
noise <- createNoise(sampleSize = tParCount, mean = 0, sd = 1)
tvMA1CoefArray <- createTVMA1CoefArray(coefFunction = sin, tParArray = tParArray)
sample <- createSample(model = createMa1, tvMA1CoefArray = tvMA1CoefArray, noise = noise)
bandwith <- 0.5
lag <- 2
nonCoverageProbability <- 0.05
meByCovHat <- computeMEbyCovHats(lag = lag, sample = sample, bandwith = bandwidth, nonCoverageProbability = nonCoverageProbability

}
