% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNonCoverageFreq.R
\name{computeNonCoverageFreq}
\alias{computeNonCoverageFreq}
\alias{}
\title{computeNonCoverageFreq}
\usage{
computeNonCoverageFreq(replicationCount, sampleSize, lag, tParArray, corArray,
  lagCount, nonCoverageProbability = 0.05, superReplicationCount,
  fileName = "")
}
\arguments{
\item{replicationCount}{a replication count.}

\item{sampleSize}{a sample size.}

\item{lag}{a lag value, true lag is lag - 1.}

\item{tParArray}{a T parameter array}

\item{corArray}{a true correlation array}

\item{nonCoverageProbability}{probability of non-coverage.}

\item{bandwidth}{a bandwidth value.}
}
\description{
For each band in bandsArray it computes isCovered and ads up all those zeros and ones. Divides this sum by replicationCount = length or dim of bandsArray
}
\examples{
myTParCount  <-2
mockTParArray  <- createTParArray(tParCount = myTParCount)
myReplicationCount <- 20
mySampleSize <-50
myLag <- 2
myLagCount <- 4
trueCorArray <- computeCor(lag = myLag,coefFunction = sin,tParArray = mockTParArray)
myBandwidth <- 0.5
myNonCoverageProbability <- 0.05
bandsBrick = createBandsBrick(sampleSize = mySampleSize,
                               tParArray  = mockTParArray,
                               lag        = myLag,
                               lagCount   = myLagCount,
                               bandwidth  = myBandwidth,
                             nonCoverageProbability = myNonCoverageProbability,
                              replicationCount       = myReplicationCount)

}
