% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNonCoverageFreqArray.R
\name{computeNonCoverageProbDist}
\alias{computeNonCoverageProbDist}
\alias{computeNonCoverageFreqArray}
\alias{}
\title{computeNonCoverageFreqDist}
\usage{
computeNonCoverageFreqArray(superReplicationCount, replicationCount, sampleSize,
  lag, tParArray, nonCoverageProbability, fileName = "")
}
\arguments{
\item{superReplicationCount}{superReplicationCount}

\item{replicationCount}{a replication count.}

\item{sampleSize}{a sample size.}

\item{lag}{a lag value, true lag is lag - 1.}

\item{tParArray}{a T parameter array}

\item{nonCoverageProbability}{probability of non-coverage.}

\item{bandwidth}{a bandwidth value.}
}
\value{
array of alpha hat
}
\description{
in a cycle for each alpha in array call NonCoverageFreqArray
}
\examples{
myTParCount  <-2
myTParArray  <- createTParArray(tParCount = myTParCount)
mySuperReplicationCount <- 4
myReplicationCount <- 4
mySampleSize <- 50
myLag <- 1
myLagCount <- 4
myNonCoverageProbability <- 0.05
nonCoverageFreqArray <- computeNonCoverageFreqArray(
superReplicationCount = mySuperReplicationCount,
replicationCount = myReplicationCount,
 sampleSize = mySampleSize,
lag = myLag,
lagCount = myLagCount,
 tParArray = myTParArray,
bandwidth = myBandwidth,
nonCoverageProbability = myNonCoverageProbability)

}
