% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBetaArray.R
\name{createBetaArray}
\alias{createBetaArray}
\title{createBetaArray}
\usage{
createBetaArray(original, rho = 0.5, gamma0)
}
\arguments{
\item{original}{A one-dimensional array created using \code{stats::arima.sim()}, passing \code{size} as the length of the array.}

\item{rho}{A \code{rho} parameter.}

\item{gamma0}{A \code{gamma0} parameter.}
}
\value{
Returning a two-dimensional array of dimensions \code{c(size - 1, size)}.
}
\description{
Creating a two-dimensional array of dimensions \code{c(size - 1, size)}.
}
\examples{
original <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
createBetaArray <- function(original, rho = 0.5)

}
