% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBetaArrayReplicated.R
\name{createBetaArrayReplicated}
\alias{createBetaArrayReplicated}
\title{createBetaArrayReplicated}
\usage{
createBetaArrayReplicated(originalReplicated, size, replicationCount, rho)
}
\arguments{
\item{originalReplicated}{A three-dimensional array of dimensions \code{c(replicationCount, size)}.}

\item{size}{The size of the one-dimensional array \code{original}.}

\item{replicationCount}{An number of replications, other words amount of times \code{createBetaArray()} is going to be executed.}

\item{rho}{A \code{rho} parameter.}
}
\value{
Returning a three-dimensional array of dimensions \code{c(size - 1, size, replicationCount)}
}
\description{
Creating a three-dimensional array of dimensions \code{c(size - 1, size, replicationCount)} using \code{createBetaArray()} as an underlying function.
}
\examples{
originalReplicated <- createOriginalMA1Replicated(replicationCount = 10, size = 10, psi = 0.5, sigma = 1)
createBetaArrayReplicated(originalReplicated, size = 10, replicationCount = 10, rho = 0.5)

}
