% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCHatMatrix.R
\name{createCHatMatrix}
\alias{createCHatMatrix}
\title{createCHatMatrix}
\usage{
createCHatMatrix(productCol, size)
}
\arguments{
\item{productCol}{A two-dimensional array of dimensions \code{c(length(original), length(original))}.}

\item{size}{The size of the one-dimensional array \code{original}.}
}
\value{
Returning a two-dimensional array of dimensions \code{c(size - 1, size - 1)}.
}
\description{
Creating a two-dimensional array of dimensions \code{c(size - 1, size - 1)}.
}
\examples{
original <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
productCol <- createProductCol(original)
createCHatMatrix(productCol, size = 10)

}
