% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCHatMatrixInALoop.R
\name{createCHatMatrixInALoop}
\alias{createCHatMatrixInALoop}
\title{createCHatMatrixInALoop}
\usage{
createCHatMatrixInALoop(productCol, size)
}
\arguments{
\item{productCol}{productCol}

\item{size}{size}
}
\value{
cHatMatrixInALoop
}
\description{
createCHatMatrixInALoop
}
\examples{

}
