% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCMatrix.R
\name{createCMatrix}
\alias{createCMatrix}
\title{createCMatrix}
\usage{
createCMatrix(size, sigma, psi)
}
\arguments{
\item{size}{The size of the one-dimensional array \code{original}..}

\item{sigma}{\code{Sigma} parameter as being the regular sigma, a.k.a. std.}

\item{psi}{he main argument passed to the model of \code{stats::arima.sim()}.}
}
\value{
Returning a two-dimensional array of dimensions \code{c(size - 1, size - 1)}
}
\description{
Creating a two-dimensional array of dimensions \code{c(size - 1, size - 1)} using a formula from the article.
}
\examples{
createCMatrix(size = 10, sigma = 1, psi = 0.5) 

}
