% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDHatMatrix.R
\name{createDHatMatrix}
\alias{createDHatMatrix}
\title{createDHatMatrix}
\usage{
createDHatMatrix(betaArray, size)
}
\arguments{
\item{betaArray}{betaArray}

\item{size}{The size of the one-dimensional array \code{original}..}
}
\value{
Returning a two-dimensional array of dimensions \code{c(size - 1, size - 1)}.
}
\description{
Creating a two-dimensional array of dimensions \code{c(size - 1, size - 1)}.
}
\examples{
original <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
betaArray <- createBetaArray <- function(original, rho = 0.5)
createDHatMatrix(betaArray, size = 10)

}
