% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDMatrix.R
\name{createDMatrix}
\alias{createDMatrix}
\title{createDMatrix}
\usage{
createDMatrix(size, sigma, psi)
}
\arguments{
\item{size}{The size of the one-dimensional array \code{original}..}

\item{sigma}{\code{Sigma} parameter as being the regular sigma, a.k.a. std.}

\item{psi}{The main argument passed to the model of \code{stats::arima.sim()}.}
}
\value{
Returning a two-dimensional array of dimensions \code{c(size - 1, size - 1)}
}
\description{
Creating a two-dimensional array of dimensions \code{c(size - 1, size - 1)} using a formula from the article.
}
\examples{
createDMatrix(size = 10, sigma = 1, psi = 0.5)

}
