% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDoubleAlphaArray.R
\name{createDoubleAlphaArray}
\alias{createDoubleAlphaArray}
\title{createDoubleAlphaArray}
\usage{
createDoubleAlphaArray(superReplicationCount, replicationCount, sampleSize, lag,
  alphaArray, tParArray, fileName = "")
}
\arguments{
\item{superReplicationCount}{how many times we produced alpha hat.}

\item{replicationCount}{how many bands we need to generate to form a single alphaHat}

\item{sampleSize}{size of the sample}

\item{lag}{lag value.}

\item{alphaArray}{a alpha parametr array}

\item{tParArray}{a T parameter array}

\item{lagCount}{how many lags we consider.}

\item{bandwidth}{a bandwidth value.}
}
\value{
A two dimensional array representing the alpha and many alpha hat for it.
}
\description{
Create double alpha hat array for every alpha in input array
}
\examples{
myAlphaCount = 4
myTParCount  <- 2
myTParArray  <- createTParArray(tParCount = myTParCount)
mySuperReplicationCount <- 3
myReplicationCount <- 6
mySampleSize <- 10
myLag <- 1
myLagCount <- computeLagCount(lag = myLag,sampleSize = mySampleSize)
myAlphaArray <- seq(from = 0.05,
                 to   = 1 - 1 / myAlphaCount,
                  by   = 1 / myAlphaCount)
nonCoverageProbabilities <- c(0.2,0.4,0.6,0.8)
myBandwidth <- 0.5
doubleAlphaArray = createDoubleAlphaArray(
superReplicationCount = mySuperReplicationCount,
 replicationCount = myReplicationCount,
sampleSize = mySampleSize,
 alphaArray =myAlphaArray,
lag = myLag,
lagCount = myLagCount,
tParArray = myTParArray,
bandwidth = myBandwidth)

}
