% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLRVof2dArray.R
\name{createLRVof2dArray}
\alias{createLRVof2dArray}
\title{createLRVof2dArray}
\usage{
createLRVof2dArray(array2d)
}
\arguments{
\item{array2d}{A \code{productCol} two-dimensional array.}
}
\value{
A scalar value of the long-run variance of a two-dimensional array.
}
\description{
Computing a long-run variance of a two-dimensional array (column-wise).
}
\examples{
original <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
productCol <- createProductCol(original)
createLRVof2dArray(productCol)

}
