% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLRVof3dArray.R
\name{createLRVof3dArray}
\alias{createLRVof3dArray}
\title{createLRVof3dArray}
\usage{
createLRVof3dArray(array3d)
}
\arguments{
\item{array3d}{A \code{productColReplicated} three-dimensional array.}
}
\value{
An array of the LRV (long-run variance) values of each column.
}
\description{
Computing a long-run variance of a three-dimensional array.
}
\examples{
originalReplicated <- createOriginalMA1Replicated (replicationCount = 10, size = 10, psi = 0.5, sigma = 1)
productColReplicated <- createProductColReplicated(originalReplicated, size = 10, replicationCount = 10)
createLRVof3dArray(productColReplicated)

}
