% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNoise.R
\name{createNoise}
\alias{createNoise}
\alias{createZ}
\title{createNoise}
\usage{
createNoise(sampleSize, mean, sd)
}
\arguments{
\item{sampleSize}{time, a.k.a. \code{N} of the sample of \code{X}.}

\item{mean}{the mean.}

\item{sd}{standard deviation.}
}
\value{
The vector of numbers.
}
\description{
Creating a normally distributed vector of \code{noise} given \code{sampleSize} (length), \code{mean} and \code{sd}.
}
\examples{
createNoise(sampleSize = 10, mean = 0, sd =  1)
createNoise(sampleSize = 20, mean = 5, sd =  10)

}
