% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOriginalMA1.R
\name{createOriginalMA1}
\alias{createOriginalMA1}
\alias{createOriginal}
\title{createOriginalMA1}
\usage{
createOriginalMA1(size, psi, sigma)
}
\arguments{
\item{size}{The size of the future one-dimensional array.}

\item{psi}{The main argument passed to the model of \code{stats::arima.sim()}.}

\item{sigma}{\code{Sigma} parameter as being the regular sigma, a.k.a. std.}
}
\value{
Returninng a one-dimensional array of size \code{size}.
}
\description{
Creating a one-dimensional array using \code{stats::arima.sim()}, passing \code{size}, \code{sigma} and \code{psi} (\code{psi} being the argument of the model).
}
\examples{
createOriginalMA1(size = 10, psi = 0.5, sigma = 1)

}
