% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProductCol.R
\name{createProductCol}
\alias{createProductCol}
\alias{createAlpha}
\title{createProductCol}
\usage{
createProductCol(original)
}
\arguments{
\item{original}{A one-dimensinal array \code{original} of length \code{size}.}
}
\value{
Returning a two-dimensional upper-triangle-like array of dimensions \code{c(length(original), length(original))}.
}
\description{
Creating a two-dimensional upper-triangle-like array of dimensions \code{c(length(original), length(original))}. Each value in a cell \code{productCol[i, i]} is a product of \code{original[i] * original[i]}.
}
\examples{
original <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
createProductCol(original)

}
