% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProductColReplicated.R
\name{createProductColReplicated}
\alias{createProductColReplicated}
\title{createProductColReplicated}
\usage{
createProductColReplicated(originalReplicated, size, replicationCount)
}
\arguments{
\item{originalReplicated}{A two-dimensional array.}

\item{size}{The size of the one-dimensional array.}

\item{replicationCount}{A number of replications, other words, an amount of times \code{createProductCol()} is going to be executed over \code{originalReplicated} .}
}
\value{
Returning a three-dimensional array of dimensions \code{c(size, size, replicationCount)}
}
\description{
Creating a three-dimensional array of dimensions \code{c(size, size, replicationCount)} using \code{createProductCol()} as an underlying function.
}
\examples{
originalReplicated <- createOriginalMA1Replicated (replicationCount = 10, size = 10, psi = 0.5, sigma = 1)
createProductColReplicated(originalReplicated, size = 10, replicationCount = 10)

}
