% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSample.R
\name{createSample}
\alias{createSample}
\alias{createX}
\title{createSample}
\usage{
createSample(sampleSize)
}
\arguments{
\item{tvMA1CoefArray}{the psi value.}
}
\value{
The vector of \code{X}
}
\description{
Creating \code{sample} given \code{tvMA1CoefArray}tvMA1CoefArray.
}
\examples{
createSample(tvMA1CoefArray = seq(from = 0, to = 1, length.out = 10))

}
