% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTParArray.R
\name{createTParArray}
\alias{createTParArray}
\alias{createSleeper}
\title{createTParArray}
\usage{
createTParArray(tParCount)
}
\arguments{
\item{tParCount}{the desired length of \code{T parameter} vector.}
}
\value{
The vector of \code{T parameter array}.
}
\description{
Create \code{T parameter array}.
}
\examples{
tParCount = 10
createTParArray(tParCount = tParCount)

}
