% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTVMA1CoefArray.R
\name{createTVMA1CoefArray}
\alias{createTVMA1CoefArray}
\alias{createPsi}
\title{createTVMA1CoefArray}
\usage{
createTVMA1CoefArray(sampleSize)
}
\arguments{
\item{tParArray}{a T parameter, a.k.a. a sleeper parameter, a sequence of numbers {0, 1}.}

\item{coefFunction}{a coeffficients function, which is an underlying function used to generate \code{ma1CoefArray} vector.}
}
\value{
The vector of the array of time variant coefficients values corresponding to T parameter values.
}
\description{
Creating \code{ma1CoefArray}, that is, the array of time variant coefficients, based on \code{tParArray} using a coeffficients function. If no coefFunction provided, \code{ma1CoefArray} vector equals to \code{tParArray} vector.
}
\examples{

createMA1CoefArray(coefFunction = sin, tParArray =  seq(from = 0, to = 1, length.out = 10))
createMA1CoefArray(coefFunction = sin, tParArray =  seq(from = 0, to = 1, by = 0.1))

}
