% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCvsCHat.R
\name{drawCvsCHat}
\alias{drawCvsCHat}
\title{drawCvsCHat}
\usage{
drawCvsCHat(cMatrix, cHatMatrix)
}
\arguments{
\item{cMatrix}{A cMatrix}

\item{cHatMatrix}{A cHatMatrix}
}
\value{
No data structures returned. Saving JPEG file located at \code{"./plots"} directory.
}
\description{
Drawing a barplot of side-by-side values of \code{cMatrix} and \code{cHatMatrix} values. Exporting it to a JPEG file saved to \code{"./plots"} directory.
}
\examples{
cMatrix <- createCMatrix(size = 10, sigma = 1, psi = 0.5) 
original <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
productCol <- createProductCol(original)
cHatMatrix <- createCHatMatrix(productCol, size = 10)
drawCvsCHat(cMatrix, cHatMatrix)

}
