% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDvsDHat.R
\name{drawDvsDHat}
\alias{drawDvsDHat}
\title{drawDvsDHat}
\usage{
drawDvsDHat(dMatrix, dHatMatrix)
}
\arguments{
\item{dMatrix}{A dMatrix}

\item{dHatMatrix}{A dHatMatrix}
}
\value{
No data structures returned. Saving JPEG file located at \code{"./plots"} directory.
}
\description{
Drawing a barplot of side-by-side values of \code{dMatrix} and \code{dHatMatrix} values. Exporting it to a JPEG file saved to \code{"./plots"} directory.
}
\examples{
dMatrix <- createDMatrix(size = 10, sigma = 1, psi = 0.5)
original <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
betaArray <- createBetaArray <- function(original, rho = 0.5)
dHatMatrix <- createDHatMatrix(betaArray, size = 10)
drawDvsDHat(dMatrix, dHatMatrix)

}
