% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawProductCol.R
\name{drawProductCol}
\alias{drawProductCol}
\title{drawProductCol}
\usage{
drawProductCol(productCol, size, mean, sigma, lagCount,
  darkestColor = "dark blue", lightestColor = "white")
}
\arguments{
\item{productCol}{A productCol two-dimensional array.}

\item{size}{The size of the one-dimensional array.}

\item{mean}{The mean.}

\item{sigma}{The sigma .}

\item{lagCount}{The lagCount .}

\item{darkestColor}{The darkest color of the canvas that is going to be used to display the tails of the \code{productCol} columns. The smaller \code{lagCount} is, the darker is the color used to display the tail.}

\item{lightestColor}{The lightest color of the canvas that is going to be used to display the tails of the \code{productCol} columns. The larger \code{lagCount} is, the lighter is the color used to display the tail.}
}
\value{
No data structures returned. Saving a JPEG file located at \code{"./plots"} directory.
}
\description{
Drawing a plot of \code{ProductCol} values. Exporting it to a JPEG file saved to \code{"./plots"} directory.
}
\examples{
original <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
productCol <- createProductCol(original)
drawProductCol(productCol, size = 10, mean = 0, sigma = 1, lagCount = 2, darkestColor = "dark blue", lightestColor = "white"

}
