% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalDifferenceKernel.R
\name{normalDifferenceKernel}
\alias{normalDifferenceKernel}
\alias{kernelGaussian4}
\title{normalDifferenceKernel}
\usage{
normalDifferenceKernel(u)
}
\arguments{
\item{u}{the length.}
}
\value{
A scalar kernel
}
\description{
A manual implementation of the 4th Gaussian kernel.

A manual implementation of the 4th Gaussian kernel.
}
\examples{
normalDifferenceKernel(10)

}
