% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveJpg.R
\name{saveJpg}
\alias{saveJpg}
\alias{saveJPG}
\title{saveJpg}
\usage{
saveJpg(fileName, path)
}
\arguments{
\item{fileName}{A string defining the path to the file that is going to be used to store the picture.}
}
\value{
No data structures returned. Saving a JPEG file located in \code{path()} directory.
}
\description{
Drawing a barplot of side-by-side values of \code{dMatrix} and \code{dHatMatrix} values. Exporting it to a JPEG file saved to \code{path()} directory.
}
\examples{

}
