% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProductCol.R
\name{saveProductCol}
\alias{saveProductCol}
\title{saveProductCol}
\usage{
saveProductCol(dataframe)
}
\arguments{
\item{dataframe}{dataframe}
}
\value{
No data structures returned. Saving a CSV file located in \code{./export} directory.
}
\description{
Saving a two-dimensional array \code{productCol} to a CSV file saved to \code{"./export"} directory.
}
\examples{
original <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
productCol <- createProductCol(original)
saveProductCol(productCol)

}
